Fix nightly test. Result of odd petsc behaviour:
This nightly test ('!TestStatic') started to fail when another test ('!TestDynamic') was added ''above'' it - turned out that !TestStatic gives
(numerically) different results depending on whether the first test, !TestDynamic, was run or not. Narrowed this down to the linear solver using max_iters=something_large in whichever is the first test run, and 1000 in the second test run. Hence when !TestStatic was the only test in the file, the linear solver would solve to convergence, whereas once !TestDynamic was added it do 1000 iters in the solve and give slightly different results. No idea why - max iters was to be set to PETSC_DEFAULT.

r15390 coincidentally sets max_iters = 1000 all the time, so now the presence of first test doesn't impact the second. Changing hardcoded-value test accordingly..





git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@15392 b5f69653-2bf5-0310-8201-b0dafb8e1405