Fix a bug with units conversion in a few model/protocol combinations (#2643).
This came up with the Grandi model and ICaL, where we units convert extracellular calcium.
The model has a top level interface component through which connections are made to the actual variable, so converting just at the boundary of the protocol component we add is insufficient - we need potentially to convert at all new connections.

git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@23831 b5f69653-2bf5-0310-8201-b0dafb8e1405