Support units converting within default expressions added by an FC protocol.
This came up with the new late sodium protocol we are trying out.  If the protocol adds an
expression with some terms using model units (since they're in the model) and some using protocol
units (since they're not) you end up needing to convert within the expression, not just at
assignments, so UnitsSet objects need to hold units in differing dimensions.

git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@25676 b5f69653-2bf5-0310-8201-b0dafb8e1405